minidlna: update to 1.1.0
This patch updates the minidlna package in OpenWRT to 1.1.0.
The upstream developers renamed the binary of minidlna to minidlnad so
we have to make the changes accordingly to luci-minidlna.
All the path to /usr/bin/minidlna should be renamed to /usr/bin/minidlnad.

Also, upstream has changed the Makefile procedure and
all the patches have been fixed upstream so they can be deleted.